backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0

# HTTP client for API calls
httpx==0.25.2

# Authentication and security
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Email sending
secure-smtplib==0.1.1

# Date and time handling
python-dateutil==2.8.2

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production server (optional)
gunicorn==21.2.0

# Additional utilities
python-jose[cryptography]==3.3.0  # For JWT tokens (future enhancement)
slowapi==0.1.9  # For rate limiting (future enhancement)

